+++ src/osgPlugins/flt/OldMaterialPaletteRecord.h
@@ -39,7 +39,7 @@ class OldMaterialPaletteRecord : public AncillaryRecord
         virtual Record* clone() const { return new OldMaterialPaletteRecord(); }
         virtual const char* className() const { return "OldMaterialPaletteRecord"; }
         virtual int classOpcode() const { return OLD_MATERIAL_PALETTE_OP; }
-        virtual int sizeofData() const { return sizeof(SOldMaterial); }
+        virtual size_t sizeofData() const { return sizeof(SOldMaterial); }
         virtual void accept(RecordVisitor& rv) { rv.apply(*this); }
 //      virtual void traverse(RecordVisitor& rv);
